Safari-609.1.11.1/LayoutTests/css3/flexbox/shrinking-column-flexbox.html

<!DOCTYPE html>

<link href="resources/flexbox.css" rel="stylesheet">

<style>
body {
  width: 800px;
  height: 600px;
}

.green {
  background: green;
}

.blue {
  background: blue;
}

.inner {
  height: 200px;
}

.small {
  height: 100px;
}
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<script>
onload = function() {
  document.body.offsetWidth;
  document.getElementById("target").classList.add("small");

  checkLayout("body");
};
</script>

<body class="flexbox column">

<div class="flexbox column flex-one-one-auto">
  <div class="flex-one-one-auto green" data-expected-height="250" id="target">
    <div class="inner"></div>
  </div>
  <div class="flex-one-one-auto blue" data-expected-height="350">
    <div class="inner"></div>
  </div>
</div>

<div id=log></div>
